requirements.txt
fastapi
uvicorn
motor
python-dotenv
sentence-transformers
faiss-cpu
google-genai
httpx